NAPALM documentation for new permissions system
In the new permissions system, there is no longer a dcim.napalm_read_device section. There's a sentence later in the same document which implies the same thing, but since the top list is a helpful configuration checklist, it would be clearer to users if it's specified there as well.